dnsdist: fix test warnings

+++ pdns/dnsdistdist/test-delaypipe_hh.cc
@@ -25,7 +25,7 @@ BOOST_AUTO_TEST_CASE(test_object_pipe) {
 
 };
 
-std::atomic<uint64_t> done = 0;
+std::atomic<int> done = 0;
 BOOST_AUTO_TEST_CASE(test_delay_pipe_small) {
   done = 0;
   struct Work
